# Reduce memory allocation