Before
```
$ go test -bench "(Fetch|App)" -benchmem
goos: darwin
goarch: amd64
pkg: github.com/kayac/go-katsubushi
BenchmarkApp-4               	  100000	     19032 ns/op	    1297 B/op	       9 allocs/op
BenchmarkAppSock-4           	  200000	      7864 ns/op	    1296 B/op	       9 allocs/op
BenchmarkClientFetch-4       	  100000	     20694 ns/op	    4816 B/op	      14 allocs/op
BenchmarkGoMemcacheFetch-4   	  100000	     22426 ns/op	    2642 B/op	      34 allocs/op
```

After
```
$ go test -bench "(Fetch|App)" -benchmem
goos: darwin
goarch: amd64
pkg: github.com/kayac/go-katsubushi
BenchmarkApp-4               	  100000	     19451 ns/op	    1185 B/op	       8 allocs/op
BenchmarkAppSock-4           	  200000	      7589 ns/op	    1185 B/op	       8 allocs/op
BenchmarkClientFetch-4       	  100000	     18636 ns/op	     592 B/op	      11 allocs/op
BenchmarkGoMemcacheFetch-4   	  100000	     21124 ns/op	    2530 B/op	      33 allocs/op
```